+++ CHANGELOG.md
@@ -10,7 +10,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added search feature for overview page [#347](https://github.com/openkfw/TruBudget/issues/347)
-- Added backwards compatibility for future versions [#360](https://github.com/openkfw/TruBudget/issues/360)
 
 ### Changed
 
@@ -26,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+- Added backwards compatibility for future versions [#360](https://github.com/openkfw/TruBudget/issues/360)
 - If a user has permissions to view subproject details, but not the permissions to see the parent project, he/she can still view the subproject details [#282](https://github.com/openkfw/TruBudget/issues/282)
 
 <!-- ### Security -->